Move classes for audio input and output to different submodule
